Dedupliser SaksStatistikkService-konstruksjon
Trekker ut felles hjelpefunksjon lagSaksStatistikkService() for å unngå
identisk konstruksjonskode i LagreSakinfoTilBigQueryJobb og
ResendSakstatistikkJobb. Endringer i oppbygging av SaksStatistikkService
gjøres nå kun ett sted.
