+++ app/src/main/kotlin/no/nav/aap/statistikk/saksstatistikk/LagreSakinfoTilBigQueryJobb.kt
@@ -11,7 +11,6 @@ import no.nav.aap.motor.ProvidersJobbSpesifikasjon
 import no.nav.aap.statistikk.LoggingKontekst
 import no.nav.aap.statistikk.behandling.BehandlingId
 import no.nav.aap.statistikk.behandling.BehandlingRepository
-import no.nav.aap.statistikk.hendelser.BehandlingService
 import no.nav.aap.statistikk.jobber.appender.JobbAppender
 import no.nav.aap.statistikk.jobber.appender.MotorJobbAppender
 import org.slf4j.LoggerFactory
@@ -132,17 +131,7 @@ class LagreSakinfoTilBigQueryJobb : ProvidersJobbSpesifikasjon {
         repositoryProvider: RepositoryProvider,
         gatewayProvider: GatewayProvider
     ): JobbUtfører {
-        val behandlingService = BehandlingService(repositoryProvider, gatewayProvider)
-        val sakStatistikkService = SaksStatistikkService(
-            behandlingService = behandlingService,
-            sakstatistikkRepository = repositoryProvider.provide(),
-            bqBehandlingMapper = BQBehandlingMapper(
-                behandlingService = behandlingService,
-                rettighetstypeperiodeRepository = repositoryProvider.provide(),
-                oppgaveRepository = repositoryProvider.provide(),
-                sakstatistikkEventSourcing = SakstatistikkEventSourcing(),
-            ),
-        )
+        val sakStatistikkService = lagSaksStatistikkService(repositoryProvider, gatewayProvider)
         return LagreSakinfoTilBigQueryJobbUtfører(
             sakStatistikkService,
             MotorJobbAppender(),
